Extract skillscmd sub-package for Kong-skillsmith integration
Move Kong option structs, dispatch, and output formatting to
skillscmd package. skills.go now only creates Smith and delegates
to skillscmd.Commands.Run().
